requirements.txt
flask>=3.0
openai>=1.30
python-dotenv>=1.0
